fib sum

+++ 0x02-functions_nested_loops/103-fibonacci.c
@@ -13,19 +13,19 @@ int main(void)
 	long i = 0;
 	long j = 1;
 	int n = 0;
-	int sum = 0;
+	long sum = 0;
 	long fib;
 
 	while (n < 4000000)
 	{
 		fib = j + i;
-		if (fib % 2 == 0 )
+		if (fib % 2 == 0)
 			sum += fib;
 
 		i = j;
 		j = fib;
 		n += 1;
 	}
-	printf("%d\n", sum);
+	printf("%ld\n", sum);
 	return (0);
 }
